[base] misc ThreadGroupSemaphore cleanups
Comply with the C++ style guide re: no default args in virtual
functions, and size_t for container types.

Also remove unneeded `explicit` characterization of 2-arg
constructor.

Follow-up to crrev.com/c/5186974

Bug: 40930642
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5386688
Commit-Queue: Sean Maher <spvw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286100}

+++ base/task/thread_pool/thread_group_semaphore.h
@@ -95,7 +95,7 @@ class BASE_EXPORT ThreadGroupSemaphore : public ThreadGroup {
   // and is no longer considered signaled at
   // OnWorkerBecomesIdleLockRequired(). Will not exceed `workers_.size()` due to
   // being deliberately capped in`EnsureEnoughWorkersLockRequired().
-  int num_active_signals_ GUARDED_BY(lock_) = 0;
+  size_t num_active_signals_ GUARDED_BY(lock_) = 0;
 
   size_t worker_sequence_num_ GUARDED_BY(lock_) = 0;
